+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased yet]
 
+### API
+* (breaking) remove `Goal`, `View` as top level statements
+* IEP 03: allow enumeration of constructed types in theory/structure block
+
 
 ## [0.7.2] - 2021-02-16